Moved redux related bits up and out of submodules
