--- assets/js/reducers/index.js
@@ -1,1 +1,0 @@
-export {default as Bag} from './bags';
