+++ assets/jsx/app.jsx
@@ -1,7 +1,28 @@
 import React from 'react';
 import {Provider} from 'react-redux';
-import configureStore from '../js/store/configureStore';
+import {
+    createStore,
+    applyMiddleware,
+    combineReducers,
+    compose
+} from 'redux';
+import thunkMiddleware from 'redux-thunk';
+import createLogger from 'redux-logger'
+
+import {bag} from '../js/reducers';
 import Bagger from '../jsx/bagger.jsx';
+
+const loggerMiddleware = createLogger()
+
+let createStoreWithMiddleware;
+
+createStoreWithMiddleware = applyMiddleware(thunkMiddleware, loggerMiddleware)(createStore);
+
+const rootReducer = combineReducers({bag});
+
+function configureStore(initialState) {
+    return createStoreWithMiddleware(rootReducer, initialState);
+}
 
 const store = configureStore();
